created new folder day2 and add functionality
