+++ Projects/liull-4r/day-2--addFunctionalityToCalculator/calculator/README.md
@@ -1,0 +1,18 @@
+# Simple Calculator 
+
+This project is a straightforward calculator application built using React. It performs basic arithmetic operations such as addition, subtraction, multiplication,modulus and division.
+
+
+<img width="274" alt="calculator" src="https://github.com/liull-4r/100DaysOfCode/assets/141246062/7c224e6e-2cc2-4b26-9abb-cb4a5d77035c">
+
+
+# Features
+Arithmetic Operations: Perform addition, subtraction, multiplication, and division modulus.
+# Technologies Used
+React: Utilized for building the user interface and managing application state.
+
+
+
+CSS: Custom styling for the calculator layout and design.
+
+
