+++ Projects/liull-4r/day-2--addFunctionalityToCalculator/calculator/src/App.jsx
@@ -1,0 +1,13 @@
+import Calculator from "./Components/Calculator";
+
+
+function App() {
+  return (
+    <>
+      <Calculator />
+  
+    </>
+  );
+}
+
+export default App;
